+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [8.2.0] 2026-02-12
+
+### Added
+
+- [TD-7861]:
+  - DD: Fetch store information
+  - LM: Upsert relations
+
 ## [8.1.0] 2026-02-09
 
 ### Added